+++ examples/src/sync_pipeline.py
@@ -1,6 +1,12 @@
 import time
 
-from plombery import register_pipeline, task
+from pydantic import BaseModel
+
+from plombery import register_pipeline, task, get_logger
+
+
+class InputParams(BaseModel):
+    what: str
 
 
 @task
@@ -9,6 +15,7 @@ def sync_task():
     This task is not async though it shouldn't block
     the rest of the app
     """
+    get_logger().debug("Im going to sleep for 10secs")
     time.sleep(10)
 
 
@@ -16,4 +23,5 @@ register_pipeline(
     id="sync_pipeline",
     description="This pipeline contains a sync task",
     tasks=[sync_task],
+    params=InputParams,
 )